+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - **MAJOR CHANGE:** Renamed `ListingCurrencies` to `FloatCurrencies`. Both `keys` and `metal` fields are now defined as `f32`. This makes more sense as this is generally used as a container type for holding original values from responses. `FloatCurrencies` should always be converted to `Currencies` in order to perform arithmetic.
 - Some method names to fit with above change.
 - "from" methods that also take in a key price are now suffixed with "with".
-- Converting `Currencies` and `FloatCurrencies` not prints `"0 keys, 0 ref"` when currencies are empty.
+- Printing `Currencies` and `FloatCurrencies` now prints `"0 keys, 0 ref"` when currencies are empty.
+- `TryFrom<FloatCurrencies>` for `Currencies` is strict now.
+- `checked_add` and `checked_sub` now require owned values.
 
 ### Added
 - `checked_get_metal_from_float` helper method.